adding font tests
